Implemented kip compilation to ckp bytecode

+++ inc/kipVersion.h
@@ -4,6 +4,6 @@
 
 namespace kip
 {
-  DLLMODE extern const uint16_t versionMajor;
-  DLLMODE extern const uint16_t versionMinor;
+  DLLMODE extern const uint8_t versionMajor;
+  DLLMODE extern const uint8_t versionMinor;
 }
